+++ test/backend/git_handler_test.dart
@@ -8,19 +8,23 @@ import 'dart:io';
 
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
-import 'package:kidney_core/src/backend/git_cloner.dart';
+import 'package:kidney_core/src/backend/git_handler.dart';
 import 'package:path/path.dart' as path;
 
 // A mock class for the ProcessRunner function.
 class MockProcessRunner extends Mock {
-  Future<ProcessResult> call(String executable, List<String> arguments);
+  Future<ProcessResult> call(
+    String executable,
+    List<String> arguments, {
+    String? workingDirectory,
+  });
 }
 
 void main() {
   // Group all tests for GitCloner
-  group('GitCloner', () {
+  group('GitHandler', () {
     late Directory tempDir;
-    late GitCloner gitCloner;
+    late GitHandler gitHandler;
     late MockProcessRunner mockProcessRunner;
 
     // Setup before each test
@@ -30,7 +34,7 @@ void main() {
       // Initialize the mock process runner
       mockProcessRunner = MockProcessRunner();
       // Create a GitCloner instance with the injected mock process runner
-      gitCloner = GitCloner(processRunner: mockProcessRunner.call);
+      gitHandler = GitHandler(processRunner: mockProcessRunner.call);
     });
 
     // Cleanup the temporary directory after each test
@@ -61,7 +65,7 @@ void main() {
         );
 
         // Act
-        await gitCloner.cloneRepo(repoUrl, targetDirectory);
+        await gitHandler.cloneRepo(repoUrl, targetDirectory);
 
         // Assert
         // Verify that the process runner was called with the correct arguments
@@ -94,7 +98,7 @@ void main() {
 
         // Act & Assert
         expect(
-          () async => await gitCloner.cloneRepo(repoUrl, targetDirectory),
+          () async => await gitHandler.cloneRepo(repoUrl, targetDirectory),
           throwsA(
             predicate(
               (e) =>
@@ -145,7 +149,7 @@ void main() {
         );
 
         // Act
-        await gitCloner.cloneRepo(repoUrl, targetDirectory);
+        await gitHandler.cloneRepo(repoUrl, targetDirectory);
 
         // Assert
         // The parent directory should have been created
@@ -165,6 +169,56 @@ void main() {
           ),
         ).called(1);
       });
+    });
+  });
+
+  group('GitHandler.checkoutBranch', () {
+    late MockProcessRunner mockProcessRunner;
+    late GitHandler gitHandler;
+
+    setUp(() {
+      mockProcessRunner = MockProcessRunner();
+      gitHandler = GitHandler(processRunner: mockProcessRunner.call);
+    });
+
+    test('successful branch checkout', () async {
+      when(
+        () => mockProcessRunner(
+          'git',
+          ['-C', 'repoDir', 'checkout', '-b', 'feature'],
+        ),
+      ).thenAnswer((_) async => ProcessResult(1, 0, '', ''));
+
+      await gitHandler.checkoutBranch('feature', 'repoDir');
+
+      verify(
+        () => mockProcessRunner(
+          'git',
+          ['-C', 'repoDir', 'checkout', '-b', 'feature'],
+        ),
+      ).called(1);
+    });
+
+    test('throws when checkout branch fails', () async {
+      when(
+        () => mockProcessRunner(
+          'git',
+          ['-C', 'repoDir', 'checkout', '-b', 'bug'],
+        ),
+      ).thenAnswer((_) async => ProcessResult(2, 1, '', 'err message'));
+
+      expect(
+        () => gitHandler.checkoutBranch('bug', 'repoDir'),
+        throwsA(
+          predicate(
+            (e) =>
+                e is Exception &&
+                e.toString() ==
+                    'Exception: Failed to checkout branch bug in repoDir: '
+                        'err message',
+          ),
+        ),
+      );
     });
   });
 }